backend/worker_service/requirements.txt
redis==7.1.0
pytest==9.0.2
pytest-cov==7.0.0
pika==1.3.2
numpy==1.26.4
pillow==12.1.0
opencv-python==4.9.0.80